requirements.txt
# Core dependencies for Chunking & Embedding Service
click>=8.0.0
openai>=1.0.0
chromadb>=0.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
jsonschema>=4.0.0

# Development and testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Additional utilities
PyYAML>=6.0.0
rich>=13.0.0  # For better CLI output formatting